refactor: put ShareFeedbackCard into PricingPageComponents
